+++ crates/sui-macros/proc-macros/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "sui-proc-macros"
+version = "0.7.0"
+authors = ["Mysten Labs <build@mystenlabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+proc-macro2 = "1"
+quote = "1"
+syn = "1.0.104"
+workspace-hack.workspace = true
+
+[target.'cfg(msim)'.dependencies]
+msim-macros = { git = "https://github.com/MystenLabs/mysten-sim.git", rev = "465f7d6c351e83003da9c9b77e8e3e352be7621e", package = "msim-macros" }
